ref: rename get_header.rs to plural
This matches the message name and protocol type.
